+++ .dotnet/CHANGELOG.md
@@ -17,6 +17,11 @@
 - Renamed properties `MaxPromptTokens` to `MaxInputTokenCount` and `MaxCompletionTokens` to `MaxOutputTokenCount` in the `ThreadRun`, `RunCreationOptions`, and `RunIncompleteReason` types. (commit_id)
 - Removed the `virtual` keyword from the `Pipeline` property across all clients. (commit_id)
 - Renamed the `Granularities` property of `AudioTranscriptionOptions` to `TimestampGranularities`. (commit_id)
+- Changed `AudioTranscriptionFormat` from an enum to an "extensible enum". (commit_id)
+- Changed `AudioTranslationFormat` from an enum to an "extensible enum". (commit_id)
+- Changed `GenerateImageFormat` from an enum to an "extensible enum". (commit_id)
+- Changed `GeneratedImageQuality` from an enum to an "extensible enum". (commit_id)
+- Changed `GeneratedImageStyle` from an enum to an "extensible enum". (commit_id)
 
 ### Bugs Fixed